+++ docs/scripts/update_imgs_and_url.jl
@@ -14,22 +14,22 @@ Pkg.activate("test")
 
 depgraph_image("Unitful", "docs/img/", fmt=:png, open=false)
 depgraph_image("Unitful", "docs/img/", fmt=:svg, open=false)
-depgraph_image("Test", "docs/img/", fmt=:svg, open=false)
+depgraph_image("Test", "docs/img/", fmt=:svg, faded=nothing, open=false)
 
-f = "ReadMe.md"
-md = read(f, String)
+readme = "ReadMe.md"
+md = read(readme, String)
 current = r"^\[dotlink\]: (.*$)"m
 m = match(current, md)
-isnothing(m) && error("[$current] not found in [$f]")
+isnothing(m) && error("[$current] not found in [$readme]")
 current_url = m.captures[1]
 
-new = PkgGraph.url("Unitful")
+new = PkgGraph.url(PkgGraph.depgraph_as_dotstr("Unitful"))
 
 if current_url == new
-    @info "Current url in [$f] still up to date"
+    @info "Current url in [$readme] still up to date"
 else
     replacement = "[dotlink]: " * new
     md = replace(md, current => replacement)
-    write(f, md)
-    println("Replaced URL in [$f]")
+    write(readme, md)
+    println("Replaced URL in [$readme]")
 end