+++ Changelog.md
@@ -33,14 +33,25 @@ per version: [Added, Changed, Fixed, Removed, Security,
 -------------
 
 
-<br>
+<!-- <br>
 
-<!-- ## v0.6  &nbsp;<sub>[![][unreleased-badge]][devlink]</sub>
+## v0.7  &nbsp;<sub>[![][unreleased-badge]][devlink]</sub>
 
 [unreleased-badge]: https://img.shields.io/badge/Unreleased-orange
 [devlink]: https://github.com/tfiers/PkgGraph.jl#development
 
 _{no changes yet}_ -->
+
+
+<br>
+
+## v0.6  &nbsp;<sub>[![][unreleased-badge]][devlink]</sub>
+
+[unreleased-badge]: https://img.shields.io/badge/Unreleased-orange
+[devlink]: https://github.com/tfiers/PkgGraph.jl#development
+
+- New keyword `faded` for `to_dot_str`.
+  Packages in the standard library are by default faded out.
 
 
 <br>